e-love-ai-service/pyproject.toml
[tool.poetry]
name = "e-love-ai-service"
version = "0.1.0"
description = ""
authors = ["paketoff <kekisreal228@gmail.com>"]
# readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.1"
pandas = "^2.2.3"
torch = "^2.5.1"
sentence-transformers = "^3.3.1"
scikit-learn = "^1.6.0"
transformers = "^4.47.1"
faker = "^33.1.0"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
pytest = "^8.3.4"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
isort = "^5.13.2"
pydantic-settings = "^2.7.1"
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
